Extract class and instance types
